Replace uses of pacmd with pactl
Reimplement the script using pactl in place of pacmd, because its list command
has a "short" option, which is easier for a script such as this to handle.

This also changes the way the verbose output is handled.  Instead of printing
the information that was collected anyway (plus the media.name), the script now
prints the entirety of the output of pactl related to the stream.