Replace a "while read" loop with piped commands
Replace the "while read" loop with two pipes: the first calculates the number of
lines of pactl output related to the loopback stream, the second calls pactl
again and uses grep with the -A option to limit the output to those lines.

This improves the speed of the script drastically.