sync_media.sh: Exclude the lost+found directory
I run sync_media.sh as a normal user, and the top-level lost+found/ directory
is owned by root, so I cannot read it.  The problem is that when rsync fails
due to the lack of permissions, it triggers an I/O error and rsync will refuse
to delete any files (the --delete option).  Passing --exclude=lost+found/ fixes
this.